requirements.txt
altgraph==0.17.2
asgiref==3.3.4
autopep8==1.6.0
certifi==2021.5.30
cryptography==3.4.7
Django==3.2.3
django-allauth==0.44.0
django-cors-headers==3.8.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
oauthlib==3.1.1
psycopg2 
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
urllib3==1.26.5
whitenoise==5.3.0
